Stabilize new resolver
With this change we mark the new resolver as stable, allowing its use on
the stable channel. We require this new version because the old one
does not necessarily determine feature sets correctly for dependencies
that are both unconditional and conditional, but with differing
features. Also see https://github.com/rust-lang/cargo/issues/9176.